Be a bit less dumb about those container ID's
So, this is great: We are using Docker which allows us to run oodles of
containers side by side. But then, we manage to always use the ID
for the first container that was started from our image.
Which means that if there are multiple jobs at once, we build multiple
things in there and introduce random races for when that container is
killed (which very likely is during one of the builds we started, so we
get failed builds).
Huzzah, let's one-line fix this.