d2d_runner: Check for return codes from DAPS instead of error messages
DAPS does not separate its messages, so that errors come out on stderr
and stdout. However, it does exit with useful return codes.

The previous approach was to use the stdout content to determine
whether there was an error. This worked ok-ish for validation because
if there is any output during a validation run during which no verbosity
is set, then there's an issue.
It did not work for building, which when successful writes the name of
the file built to stdout, and when unsuccessful writes an error. We
looked for `Stop.` among DAPS's messages but it appears GNU make no
longer emits that word at all.

We now check $? for both validation and builds and that should take care
of the issue.