d2d_runner: Add JSON variant of filelist in addition to flat filelist
The new JSON filelist (filelist.json) is off by default (-j=1 to enable)
and includes a lot more information which should also be a lot more
parseable for Docserv2.
The keys are MD5 hashes of "DC-file format", because the keys would not
be unique when building multiple formats of the same document and `jq`
would remove them during formatting.

This adds a dependency on the `jq` tool (only needed when -j=1 is
used).